mantis 35425 and 35426 : correct validator check parameters

+++ mobi.chouette.exchange.validator/src/main/java/mobi/chouette/exchange/validator/ValidateParameters.java
@@ -24,7 +24,7 @@ public class ValidateParameters extends AbstractParameter {
 	
 	@Getter @Setter
 	@XmlElement(name = "reference_ids")
-	private List<Integer> ids;
+	private List<Long> ids;
